chore: renaming utils folder to helpers
