+++ dashboard/components/inventory/hooks/useInventory/helpers/getInventoryListAndStats.ts
