+++ dashboard/components/inventory/hooks/useInventory/helpers/getInventoryListFromAFilter.ts
